+++ .github/workflows/test.yaml
@@ -19,7 +19,7 @@ jobs:
             - name: Install dependencies
               run:
                 pnpm install --frozen-lockfile
-          
+
             - name: Run test
               run:
-                pnpm build-and-test run
+                pnpm run test
