+++ package.json
@@ -2,13 +2,8 @@
   "name": "theseed-skin-buma",
   "version": "1.0.0",
   "description": "Frontend skin for the seed engine",
-  "main": "index.js",
   "scripts": {
-    "build-modular-bulma": "sass ./scss/bulma.scss ./css/bulma.css && cross-env NODE_ENV=BUILD postcss --config postcss.bulma.config.js ./css/bulma.css > ./css/bulma.min.css && rimraf ./css/bulma.css",
-    "build-layout-css": "sass ./scss/layout.scss ./css/layout.css && cross-env NODE_ENV=BUILD postcss --config postcss.bulma.config.js ./css/layout.css > ./css/layout.min.css && rimraf ./css/layout.css",
-    "build": "pnpm build-modular-bulma && pnpm build-layout-css",
-    "test": "vitest",
-    "build-and-test": "pnpm build && pnpm test"
+    "test": "vitest --run"
   },
   "repository": {
     "type": "git",
@@ -23,20 +18,12 @@
   "devDependencies": {
     "@testing-library/vue": "^8.1.0",
     "@vitejs/plugin-vue": "^5.2.4",
-    "cross-env": "^7.0.3",
-    "cssnano": "^5.1.15",
     "jsdom": "^26.1.0",
-    "postcss": "^8.5.6",
-    "postcss-cli": "^9.1.0",
-    "postcss-modules": "^4.3.1",
     "prettier": "^2.8.8",
-    "rimraf": "^3.0.2",
-    "sass": "^1.92.1",
     "vitest": "^3.2.4",
     "vue": "^3.5.21",
     "vue-final-modal": "^4.5.5",
     "vue-router": "^4.5.1",
-    "vue-template-compiler": "^2.7.16",
     "vuex": "^4.1.0"
   },
   "dependencies": {
@@ -47,5 +34,5 @@
     "@fortawesome/vue-fontawesome": "^3.1.2",
     "bulma": "1.0.3"
   },
-  "packageManager": "pnpm@10.5.2"
+  "packageManager": "pnpm@10.16.1"
 }
